removed play api, added updating results textview to postgamefragment so user can see their score being added up

+++ src/dev/emmaguy/higherorlower/HigherOrLowerGame.java
@@ -14,7 +14,7 @@ public interface HigherOrLowerGame {
     void setOnTimeRemainingChangedListener(OnTimeRemainingChanged timeRemainingChanged);
     
     public interface OnGameOver {
-	public void onGameOver(long score, long millisecondsRemaining, int leaderboardId);
+	public void onGameOver(int score, long millisecondsRemaining, int leaderboardId);
     }
     
     public interface OnCardChanged {
